refactor: improves domain structure
