--- src/components/Objective/index.ts
@@ -1,1 +1,0 @@
-export * from './types'
