+++ src/components/Pages/Objectives/objectives.tsx
@@ -3,7 +3,7 @@ import { useIntl } from 'react-intl'
 import { useSetRecoilState } from 'recoil'
 
 import PageContent from 'src/components/Base/PageContent'
-import { TeamViewCardList } from 'src/components/Company/Team/View'
+import { TeamViewCardList } from 'src/components/Team/View'
 import { pageTitleAtom } from 'src/state/recoil/page'
 
 import messages from './messages'
