+++ src/components/Team/types.ts
@@ -1,0 +1,20 @@
+import { Company } from 'src/components/Company/types'
+import { KeyResult } from 'src/components/KeyResult/types'
+import { User } from 'src/components/User/types'
+
+export interface GetRootTeamsAndCompaniesQueryData {
+  teams: Array<Partial<Team>>
+  companies: Array<Partial<Company>>
+}
+
+export interface Team {
+  id: string
+  name: string
+  description?: string
+  createdAt: Date
+  updatedAt: Date
+  company: Company
+  keyResults?: KeyResult[]
+  users?: User[]
+  teams?: Team[]
+}
