+++ src/components/User/index.ts
@@ -1,5 +1,3 @@
 export { default as NamedAvatar } from './NamedAvatar'
 
 export { default as Me } from './Me'
-
-export * from './types'
