+++ src/state/recoil/company/constants.ts
@@ -1,0 +1,1 @@
+export const PREFIX = 'COMPANY'
