+++ src/state/recoil/company/index.ts
@@ -1,0 +1,1 @@
+export { default as companyAtomFamily } from './atom-family'
