+++ src/state/recoil/key-result/selectors/build-partial-selector.ts
@@ -4,9 +4,9 @@ import setWith from 'lodash/setWith'
 import snakeCase from 'lodash/snakeCase'
 import { DefaultValue, selectorFamily } from 'recoil'
 
-import { Team } from 'src/components/Company/types'
 import { KeyResult } from 'src/components/KeyResult/types'
 import { Objective } from 'src/components/Objective/types'
+import { Team } from 'src/components/Team/types'
 import { User } from 'src/components/User/types'
 import keyResultAtomFamily from 'src/state/recoil/key-result/atom-family'
 import { RecoilSpecificationGetter, RecoilSpecificationSetter } from 'src/state/recoil/types'
