+++ src/state/recoil/key-result/selectors/key-result.ts
@@ -1,7 +1,7 @@
 import deepmerge from 'deepmerge'
 import { DefaultValue, selectorFamily } from 'recoil'
 
-import { KeyResult } from 'src/components/KeyResult'
+import { KeyResult } from 'src/components/KeyResult/types'
 import { PREFIX } from 'src/state/recoil/intl/constants'
 import keyResultAtomFamily from 'src/state/recoil/key-result/atom-family'
 import { RecoilSpecificationSetter } from 'src/state/recoil/types'
