+++ src/state/recoil/team/constants.ts
@@ -1,0 +1,1 @@
+export const PREFIX = 'TEAM'
