+++ src/state/recoil/team/index.ts
@@ -1,0 +1,1 @@
+export { default as teamAtomFamily } from './atom-family'
